feat: implement Phase 1 of Landing-PWA separation
- Add React Router DOM v7.8.2 dependency
- Create project structure: pages/, router/, hooks/ directories
- Move App.tsx content to pages/FinanceApp.tsx
- Create basic AppRouter with / and /app routes
- Update main.tsx to use router system
- Add temporary landing page for testing

Phase 1 complete: Basic routing infrastructure ready
Next: PWA detection and guard components
